+++ src/components/pages/index.ts
@@ -1,5 +1,6 @@
 export * from './TVDashboard/TVDashboard';
 export * from './HomePage';
+export * from './DevCornerOneDashboard';
 export * from './DevCornerTwoDashboard';
 export * from './ConferenceRoomDashboard';
 export * from './TrevorDashboard';